+++ client/src/components/forms/index.js
@@ -14,4 +14,5 @@ export {
   UnsavedChangesModal,
   DeleteModal,
   RenderForm,
+  StatefulTabs,
 } from "./utils";