+++ client/src/components/forms/utils/FormTemplate.js
@@ -1,7 +1,6 @@
 import React from "react";
 import { contains, kebabCase } from "lazy-z";
 import PropTypes from "prop-types";
-import ToggleForm from "./ToggleForm";
 import DynamicFormModal from "./DynamicFormModal";
 import { RenderForm } from "./StatelessFormWrapper";
 import StatefulTabs from "./StatefulTabs";