+++ client/src/components/pages/cloud-services/CloudServices.js
@@ -39,9 +39,9 @@ import {
   CraigFormGroup,
   CraigToggleForm,
   DynamicFormModal,
+  StatefulTabs,
 } from "../../forms/utils";
 import DynamicForm from "../../forms/DynamicForm";
-import StatefulTabs from "../../forms/utils/StatefulTabs";
 import { craigForms } from "../CraigForms";
 import { getServices } from "../../../lib/forms/overview";
 import { docTabs } from "../diagrams/DocTabs";